backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
langchain==0.3.13
langchain-groq==0.2.1
langchain-openai==0.2.14
langchain-core==0.3.28
langchain-community==0.3.13
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
